Split nnvm part
